requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
pydantic==2.5.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
langdetect==1.0.9
chardet==5.2.0
aiofiles==23.2.1
bcrypt==4.1.2
python-dotenv==1.0.0
